requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.13.0=pypi_0
astunparse=1.6.3=pypi_0
ca-certificates=2021.5.30=h5b45459_0
cached-property=1.5.2=pypi_0
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py37h03978a9_0
charset-normalizer=2.0.4=pypi_0
cycler=0.10.0=pypi_0
et_xmlfile=1.0.1=py_1001
flatbuffers=1.12=pypi_0
gast=0.3.3=pypi_0
google-auth=1.34.0=pypi_0
google-auth-oauthlib=0.4.5=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.32.0=pypi_0
h5py=2.10.0=pypi_0
idna=3.2=pypi_0
importlib-metadata=4.6.3=pypi_0
jdcal=1.4.1=py_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=pypi_0
markdown=3.3.4=pypi_0
matplotlib=3.4.2=pypi_0
numpy=1.19.3=pypi_0
oauthlib=3.1.1=pypi_0
openpyxl=3.0.7=pyhd8ed1ab_0
openssl=1.1.1k=h8ffe710_0
opt-einsum=3.3.0=pypi_0
pandas=1.3.1=pypi_0
pillow=8.3.1=pypi_0
pip=21.2.2=pyhd8ed1ab_0
protobuf=3.17.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=2.4.7=pypi_0
python=3.7.10=h7840368_100_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.7=2_cp37m
pytz=2021.1=pypi_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scipy=1.7.1=pypi_0
setuptools=49.6.0=py37h03978a9_3
six=1.15.0=pypi_0
sqlite=3.36.0=h8ffe710_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow-addons=0.13.0=pypi_0
tensorflow-estimator=2.4.0=pypi_0
tensorflow-gpu=2.4.0=pypi_0
termcolor=1.1.0=pypi_0
tsaug=0.2.1=pypi_0
typeguard=2.12.1=pypi_0
typing-extensions=3.7.4.3=pypi_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.6=pypi_0
vc=14.2=hb210afc_5
vs2015_runtime=14.29.30037=h902a5da_5
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3deb0d_0
wincertstore=0.2=py37h03978a9_1006
wrapt=1.12.1=pypi_0
zipp=3.5.0=pypi_0
